Bug 1187076 - Warn at end of SPS buffers. r=jya
Previously we asserted if we tried to read past the end
of a buffer in parsing H.264 SPS headers. This only affected
debug builds, but the failures were reported as crashes in
automated testing.

Therefore, commute the assert to a warning and silently accept
bad data of this type. The read is safe with the first assert
removed because of the subsequent check, and a similar one in
the parent class.

UltraBlame original commit: c75f9ca74a29c05a3457d33531548c63f9b6d784