Bug 1195436 - Fuzz scrollbar reftest failures with APZ enabled. r=mstange
In this case the test image is painted using -moz-element which layerizes
differently than not using -moz-element when APZ is enabled. This slight
layerization difference causes a few pixels to be different and the reftest
to fail. Fuzzing it seems reasonable as the fundamental nature of the test
is unchanged.

UltraBlame original commit: 63cc176d76a66b1807ea4b9500651063032123b1