Bug 1203074: Move caret to end of textual input field before sending keys
When a TEXTAREA element is focused it returns the cursor to the last
position was at, or places it last.  INPUT type="text" (or any other
textual input element) places the caret at the beginning.  Because of
this we move the caret to the end of the input field.  The next time
the element is focussed, the cursor should move to the end.

The layout touch caret tests relied on the caret being left in its
previous position.  This patch addresses that by using the advanced user
interaction API for these test cases.

r=jgriffin

UltraBlame original commit: 86bd982df1f9b339dd94d3dac25cbba7d5f20a15