Bug 1193215 -  Support for passing test directories through mach try, r=chmanchester
This adds support for web-platform-tests to mach try. It changes the implementation
so that instead of passing paths to manifests, the user passes arbitary paths in the
source tree, and tests under that path are run, with test discovery mainly left to
the harness.

UltraBlame original commit: 067d9325416abe5740fa5149e69344b02c41d69f