Bug 1211591 - Part 1: Remove incorrect sql statement from nsPermissionManager::RemovePermissionsForApp, r=jdm
This initial query to clear the database was initially removed because of the removal of the appId column, and replaced with the following logic which parses the origin strings to obtain the appId property, and then removes those entries using `AddInternal`. Unfortunately, during a rebase, it appears as though the code crept back into the codebase.

I'm not sure why this codepath wasn't ever tested before bkelly noticed it, but we should probably add a test of some form which actually runs the RemovePermissionsForApp codepath, to ensure that it actually works.

UltraBlame original commit: 0b75225d75f73cd6a7f0b7e9afc8646f09a6432b